cleans up directory after first component commit and updates readme to better outline current state of the project + next steps.
