+++ project/requirements.txt
@@ -1,0 +1,5 @@
+Flask==0.10.1
+Jinja2==2.7.3
+MarkupSafe==0.23
+Werkzeug==0.10.1
+itsdangerous==0.24
