--- project/templates/poll.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title> Polling Site 1.0 </title>
-  </head>
-  <body>
-    <script type="javascript/"
-    <div> Hello, World! </div>
-    <div> Survey form goes here </div>
-  </body>
- </html>
